separate mba tests
